More ANSI-like SUBSTITUTE[-IF[-NOT]]
Equally naïve  as the SUBSTITUTE  implementation, but useful  enough for
many cases.

If    merged,   issue    https://github.com/jscl-project/jscl/issues/239
could/should likely be expanded to reference this fact as well.

At  this point,  support  is added  for :count,  :start,  and :end,  but
notably not :from-end, which WARNs and returns nil.

There is likely  a case to be  made for an optimized  version that works
through the relatively fast regexp engine in the Javascript compiler for
the most common case of (substitute char₁ char₂ string) only.

This  branch currently  fails to  pass the  tests (actually,  it crashes
run-tests.sh).