Fix up DIGIT-CHAR-P in Boot
I am assuming that the reason digit-char-p / digit-char are defined in boot.lisp
is because they're needed before char.lisp can be loaded; but that does not mean
that we want to have two copies with different capabilities. It seems that the
latter versions should be usable in boot.lisp.