DIGIT-CHAR-P assertions.
This function has been annoyingly in the middle of the various
radix-related work on FORMAT and the READer and code generator;
so this patch adds an insane level of assertions to it, to
try to make it more obvious when Things Go Wrong.

These should probably be rolled-back later, but are useful for
the moment.