Support for radix, Unicode
Bundled set of changes to support reading and printing
things in arbitrary numeric radix 2-36;

 #\U+xxxx character literals
 #\b #\o #\x and #\NNr numeric literals
 char-name/name-char "U+xxxx" names
 #\Page #\Backspace #\Rubout #\Return

PRINC/FORMAT ~A corrections in a few cases

Partial to reasonable support in FORMAT for:
~a ~s ~w
~d ~x ~NNr
~% ~& ~|

Also bundled in are fixes for SUBSTITUTE[-IF[-NOT]]
and EVERY, as these were dependencies.

*read-base* and *print-base* support
PARSE-INTEGER

Correction to Javascript in CHAR-UPCASE in edge cases

Misc fixes discovered during the implementation
of the above.