revert to JSCL::!MACROEXPAND-1
Something in src/numbers.lisp is not working out properly with
the SBCL MACROEXPAND-1, and somehow, it does work with the
JSCL implementation, so for now, reverting.

Yes, this is hoodoo for the sake of pragmatism. I imagine that
there is some bug in ANSI-compliance of the macroexpander
that will, when fixed, break
src/numbers.lisp again.