Cross-compilation Quasiquote handling
When cross-compiling,  SBCL's reader can leave  quasiquotes (backquotes)
in its own  internal form. This defines the backquote  as though it were
a   macro,  and   binds   it   to  an   acceptable   (but  not   public)
SB-IMPL::EXPAND-QUASIQUOTE    to    handle     rewriting    the    form.
There's a possibility of some weirdness with this, but cross-compilation
tends to be weird some days.